Split up service tests
